Code cleaning

+++ src/main/java/ingsw/utilities/ControllerTimer.java
@@ -25,11 +25,11 @@ public class ControllerTimer {
     }
 
     public void startLoginTimer(int loginSeconds, Controller controller, boolean hasStarted) {
-        timer.schedule(new LaunchMatch(controller, hasStarted), loginSeconds * 1000);
+        timer.schedule(new LaunchMatch(controller, hasStarted), (long) loginSeconds * 1000);
     }
 
     public void startTurnTimer(int turnSeconds, GameManager gameManager) {
-        timer.schedule(new EndTurn(gameManager), turnSeconds * 1000);
+        timer.schedule(new EndTurn(gameManager), (long) turnSeconds * 1000);
     }
 
     public void cancelTimer() {
